requirements.txt
fastapi
uvicorn
langgraph
langchain
groq
pymupdf
python-dotenv
python-multipart
pytest
pytesseract 
pillow